% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_age.R
\name{gen_apache_age}
\alias{gen_apache_age}
\title{Generates the APACHE Age}
\usage{
gen_apache_age(dt, window = c(0, 24))
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}
}
\description{
Generates the APACHE age score;
}
\examples{
ddata <- NULL

ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[ , ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[ , (age) := sample(seq(0,99,1), 200, replace = T)]
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
system.time(gen_apache_age(ddata, window = c(0,24)))
ddata[time \%between\% c(0,24), .N, by = c("site","episode_id", "apache_age")]

}

