% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_aki.R
\name{gen_apache_aki}
\alias{gen_apache_aki}
\title{Generates the APACHE Acute Kidney Injury score}
\usage{
gen_apache_aki(dt, crrt, window, format = "dataItem")
}
\arguments{
\item{dt}{data.table containing physiological data}

\item{crrt}{Numeric.Binary vector of 0 or 1. Assessment of prior chronic renal insufficiency.}

\item{window}{Numeric.Vector delimiting boundaries for time-window.}

\item{format}{String. The format chosen for data items. Could be "dataItem", "shortName" or "NHICcode".
See relabel_cols for more informations.}
}
\examples{
ddata <- NULL
ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
ddata[, ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[, ("Creatinine") := sample(seq(20,1000,11), 200, replace = T)]
ddata[, crrt := sample(c(0,1), 200, replace = T)]
system.time(gen_apache_aki(ddata, crrt = ddata[, crrt], window = c(0,24), format = "dataItem"))
ddata[time \%between\% c(0,24), .N, by = c("site","episode_id", "apache_aki")]

}

