% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_chronic.R
\name{gen_apache_chronic}
\alias{gen_apache_chronic}
\title{Generates the APACHE Chronic Health score}
\usage{
gen_apache_chronic(dt)
}
\arguments{
\item{dt}{data.table containing physiology data}
}
\description{
Generates the APACHE Chronic Health score; requires 15 comorbidity variables as inputs. A warning is
displayed in case of missing fields
\describe{This function requires the output of the helper function gen_comorbidity, and transform it into binary variable}
\describe{The function needs two fields into the datatable, related to the type of admission and the classification of surgery.
Items have to be coded by the Yaml dictionnary as follow: }
\describe{- Admission}{
   \itemize{
     \item L: unplanned Local admission ;
     \item U: Unplanned transfer in ;
     \item P: Planned transfer in ;
     \item S: planned local Surgical admission ;
     \item M: planned local Medical admission ;
     \item R: Repatriation.}}
 \describe{- Surgery}{
   \itemize{
     \item M: Emergency ;
     \item U: Urgent ;
     \item S: Scheduled ;
     \item L: eLective}}
}
\examples{
dt <- NULL
dt$"Admission type" <- c(sample(c("L", "U", "P", "S", "M", "R"), 200, replace = T))
dt$"classification of surgery" <- c(sample(c("M", "U", "S", "L"), 200, replace = T))
dt$"d_comorbidity" <- c(sample(c(0, 1, 2, 3, 4, NA), 200, replace = T))
dt <- as.data.table(dt)
gen_apache_chronic(dt)
dt[, .N, by = b_comorbidity]
dt[, .N, by = apache_chronic]

}

