% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_gcs.R
\name{gen_apache_gcs}
\alias{gen_apache_gcs}
\title{Generates the APACHE Glasgow Coma Scale score}
\usage{
gen_apache_gcs(dt, window, format = "dataItem")
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}

\item{format}{String. The format chosen for data items. Could be "dataItem", "shortName" or "NHICcode".
See relabel_cols for more informations.}
}
\description{
Generates the APACHE gcs score;
}
\examples{
ddata <- NULL
gcs <- "GCS - total"
ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[ , ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[ , (gcs) := sample(seq(3,15,1), 200, replace = T)]
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
system.time(gen_apache_gcs(ddata, window = c(0,24), format = "dataItem"))
ddata[time \%between\% c(0,24), .N, by = c("site","episode_id", "apache_gcs")]

}

