% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_map.R
\name{gen_apache_map}
\alias{gen_apache_map}
\title{Generates the APACHE Arterial Pressure score}
\usage{
gen_apache_map(dt, window = c(0, 24))
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}
}
\description{
Generates the APACHE Arterial Pressure score;
}
\examples{
ddata <- NULL
map <- "Mean arterial blood pressure - Art BPMean arterial blood pressure"
ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
ddata[, ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[, (map) := sample(seq(30,120,1), 200, replace = T)]
system.time(gen_apache_map(ddata, window = c(0,24)))
ddata[time \%between\% c(0,24), .N, by = c("site","episode_id", "apache_map")]

}

