% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_pred.R
\name{gen_apache_pred}
\alias{gen_apache_pred}
\title{Generates the predicted mortality frome the APACHE score}
\usage{
gen_apache_pred(dt, window, imput = NULL)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}
}
\description{
Generates the predicted mortality using the formula given by Knaus et al.
\describe{This function needs:}{
   \itemize{
     \item The results of the apache_score function
     \item A weight for the primary diagnostic for admission
     \item A Medical or Post-operative condition for admission}}
}
\examples{
ddata <- NULL
ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[, ("apache_score") := sample(seq(0,60,1), 200, replace = T)]
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
ddata[, ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[, ("weight") := sample(seq(-0.5, 1.75, 0.01), 200, replace = T), by = c("site", "episode_id")]
ddata[, ("emergent_surgery") := sample(c("Yes", "No"), 200, replace = T), by = c("site", "episode_id")]
gen_apache_pred(ddata, window = c(0,24))
ddata[time \%between\% c(0,24), .N, by = c("site", "episode_id", "risk")]

}

