% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_rf.R
\name{gen_apache_rf}
\alias{gen_apache_rf}
\title{Generates the APACHE Respiratory Failure score}
\usage{
gen_apache_rf(dt, window, format = "dataItem")
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}

\item{format}{String. The format chosen for data items. Could be "dataItem", "shortName" or "NHICcode".
See relabel_cols for more informations.}
}
\description{
Generates the APACHE Respiratory Failure score;
\describe{ This function use the results of gen_grad to calculate the APACHE score for respiratory failure component}
}
\examples{
ddata <- NULL
ddata$"grad_rf" <- c(sample(seq(10,50,0.1), 200, replace = T))
ddata <- as.data.table(ddata)
ddata[,("Inspired fraction of oxygen") := sample(seq(0.21,1,0.01), 200, replace = T)]
ddata[, ("time") := sample(seq(1,72,1), 200, replace = T)]
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
ddata[, ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[, ("PaO2/FiO2 ratio") := sample(seq(4,40,1), 200, replace = T)]
ddata[, ("PaO2 - ABG") := sample(seq(4,90,1), 200, replace = T)]
gen_apache_rf(ddata, window = c(0,24), format = "dataItem")
ddata[time \%between\% c(0,24), .N, by = c("site", "episode_id", "apache_rf")]

}

