% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_rr.R
\name{gen_apache_rr}
\alias{gen_apache_rr}
\title{Generates the APACHE Respiratory Rate score}
\usage{
gen_apache_rr(dt, window, format = "dataItem")
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}

\item{format}{String. The format chosen for data items. Could be "dataItem", "shortName" or "NHICcode".
See relabel_cols for more informations.}
}
\description{
Generates the APACHE Respiratory Rate score. This function needs the result of the gen_q_rr helper function.
}
\examples{
ddata <- NULL
ddata$time <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[,("q_rr") := sample(seq(6,60,1), 200, replace = T)]
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
ddata[, ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
gen_apache_rr(ddata, window = c(0,24), format = "dataItem")
ddata[time \%between\% c(0,24), .N, by = c("site", "episode_id", "apache_rr")]

}

