% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_score.R
\name{gen_apache_score}
\alias{gen_apache_score}
\title{Generates the APACHE Score}
\usage{
gen_apache_score(dt, window)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}
}
\description{
Generates the APACHE score from the calculated variables;
}
\examples{
ddata <- NULL
ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[, ("apache_map") := sample(seq(1,4,1), 200, replace = T)]
ddata[, ("apache_hr") := sample(seq(1,4,1), 200, replace = T)]
ddata[, ("apache_ht") := sample(seq(1,4,1), 200, replace = T)]
ddata[, ("apache_rr") := sample(seq(1,4,1), 200, replace = T)]
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
ddata[, ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
gen_apache_score(ddata, window = c(0,24))
ddata[time \%between\% c(0,24), .N, by = c("site","episode_id", "apache_score")]


}

