% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_temp.R
\name{gen_apache_temp}
\alias{gen_apache_temp}
\title{Generates the APACHE Temperature score}
\usage{
gen_apache_temp(dt, window = c(0, 24))
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}
}
\description{
Generates the APACHE Temperature score;
}
\examples{
ddata <- NULL
temperature <- "Temperature - Central"
ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[ , ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[ , (temperature) := sample(seq(35,45,0.1), 200, replace = T)]
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
system.time(gen_apache_temp(ddata, window = c(0,24)))
ddata[time \%between\% c(0,24), .N, by = c("site","episode_id", "apache_temp")]

}

