% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_wbc.R
\name{gen_apache_wbc}
\alias{gen_apache_wbc}
\title{Generates the APACHE White Blood Cell Count score}
\usage{
gen_apache_wbc(dt, window = c(0, 24))
}
\arguments{
\item{dt}{data.table containing physiological data}

\item{window}{Numerical.Vector delimiting boundaries for time-window.}
}
\description{
Generates the APACHE wbc score;
}
\examples{
ddata <- NULL
ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
ddata[, ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[, ("White cell count") := sample(seq(2,50,0.1), 200, replace = T)]
system.time(gen_apache_wbc(ddata, window = c(0,24)))
ddata[time \%between\% c(0,24), .N, by = c("site","episode_id", "apache_wbc")]

}

