% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_comorbidity}
\alias{gen_comorbidity}
\title{Generate Comorbidity APACHE Score}
\usage{
gen_comorbidity(dt, fields)
}
\arguments{
\item{fields.}{Vector of strings naming the fields requested for the comorbidity score calculation.
Fields have to be formatted according to ITEM_REF yaml file, as "NHICcode", or "dataItem", or "shortName".}
}
\description{
@description gen_comorbidity counts the number of comorbidity available in dataset. If missing data, gen_comorbidity imputes missing value as 0.
 \describe{A function that counts the number of comorbidity expressed, according to APACHE definitions:}{
 \describe{- Chronic Heart Failure :}{
   \itemize{
     \item NYHA 4}}
 \describe{- Hepatic Disease :}{
   \itemize{
     \item Biopsy Proven Cirrhosis ;
     \item Documented Portal Hypertension ;
     \item Upper Gastro-Intestinal Bleedings due to Portal Hypertension ;
     \item Hepatic encephalopathy or comatose episode.}}
 \describe{- Chroncic respiratory insufficiency:}{
   \itemize{
     \item Restrictive or Obstructive or Vascular Disease with severe impairement of physical activities ;
     \item Secondary Polycythemia ;
     \item Documented Chronic Hypoxia or Hypercarbia ;
     \item Severe Pulmonary Hypertension or Long-Term Oxygenotherapy.}}
 \describe{- Chronic Renal Replacement Therapy}
 \describe{- Imunosuppression :}{
   \itemize{
     \item Immunosuppressive Therapy ;
     \item Chemotherapy ;
     \item Radiotherapy ;
     \item Long Lasting Corticotherapy ;
     \item Oncohaematologic Disease ;
     \item AIDS.}}}
}
\examples{
# library(data.table)
# ddata <- NULL
# ddata$crrt <- sample(c(1,0),200, replace = T)
# ddata$chemotherapy <- sample(c(1,0),200, replace = T)
# ddata <- as.data.table(ddata)
# ddata <- gen_comorbidity(dt = ddata, format = "dataItem")
}

