% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_grad}
\alias{gen_grad}
\title{Generate Alveolar to Arterial gradient}
\usage{
gen_grad(dt, format = "dataItem")
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{format}{String. The format chosen for data items. Could be "dataItem", "shortName" or "NHICcode".
See relabel_cols for more informations.}
}
\description{
gen_grad derives an Alveolar to Arterial gradient from fio2 and/or PaO2 and PaCO2 fields
\describe{Requires FiO2 and/or P/F ratio and PaO2 formatted as dataItem, or NHICcode, or shortName, among:}{
   \itemize{
       \item Inspired fraction of oxygen / NIHR_HIC_ICU_0150 / fiO2
       \item PaO2/FiO2 ratio / NIHR_HIC_ICU_0913 / pf_ratio}}
\describe{Calculations are performed according to the following Formula:}{
    \itemize{
       \item (A-a)O2 Gradient = (( PAtm - PH2O) * FIO2 - ( PaCO2 / RQ )) - PaO2
       \item with RQ = 1 , PAtm = 100, PH2O = 6.2 ; (A-a)O2 Gradient = (100 - 6.2) * FIO2 - PaCO2 - Pa02}}
}
\examples{
ddata <- NULL
ddata$"Inspired fraction of oxygen" <- sample(seq(0,1,0.1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[, ("PaO2 - ABG") := sample(seq(4,8,0.1), 200, replace = T)]
ddata[, ("PaCO2 - ABG") := sample(seq(3,6,0.1), 200, replace = T)]
ddata[, ("PaO2/FiO2 ratio") := sample(seq(4,40,1), 200, replace = T)]
gen_grad(ddata, format = "dataItem")
ddata[, .N, grad_rf]

}

