% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_map}
\alias{gen_map}
\title{Generate MAP from blood pressure}
\usage{
gen_map(bps, bpd)
}
\arguments{
\item{bps}{Numerical. Systolic blood pressure.}

\item{bpd}{Numerical. Diastolic blood pressure.}
}
\description{
gen_map derives MAP from systolic and diastolic blood pressure
}
\examples{
# ddata <- NULL
# ddata$bps <- as.data.table(sample(seq(60, 200, 1), 200, replace = T))
# ddata$bpd <- sample(seq(20, 120, 1), 200, replace = T)
#system.time(ddata$map <- gen_map(ddata$bps, ddata$bpd))
}

