% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_ppv}
\alias{gen_ppv}
\title{Various helper functions that GENerate derived variables}
\usage{
gen_ppv(dt, t_ = time, id_ = id, rrate_ppv_)
}
\arguments{
\item{dt}{data.table containing physiological data}

\item{t_}{Numeric. time}

\item{id_}{Numeric. Unique identifier for patient}
}
\description{
gen_ppv creates a logical vector if positive pressure ventilated
}
\examples{
# gen_ppv(ddata, time, id, Total_Resp_Rate_Ventil)
# gen_ppv makes a logical vector if positive pressure ventilated
}

