% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_smr.R
\name{gen_smr}
\alias{gen_smr}
\title{Generates the APACHE White Blood Cell Count score}
\usage{
gen_smr(dt, pred, observed, group)
}
\arguments{
\item{dt}{data.table containing physiology data}

\item{format}{Strings. The format that have been chosen by the users at naming the fields of the datatable.
see relabelcols for more informations.}
}
\description{
Generates the APACHE wbc score;
}
\examples{
# system.time(gen_apache_wbc(ddata, format = "dataItem"))
# table(ddata$apache_wbc, useNA="always")
}

