% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_weights}
\alias{gen_weights}
\title{Generate weights variable from the primary diagnosis for asmission
gen_weights derives a Haematocrit from haemoglobin and mean globular volume}
\usage{
gen_weights(dt)
}
\arguments{
\item{dt}{data for computation.}
}
\description{
Generate weights variable from the primary diagnosis for asmission
gen_weights derives a Haematocrit from haemoglobin and mean globular volume
}

