% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{reduce}
\alias{reduce}
\title{Generate 1d datatable summarizing 2d variables
reduce allows to reduce datatable according to time-window selected and according to
the iteration of values for selected variables, into a 1d table with one row by admission}
\usage{
reduce(dt, window)
}
\arguments{
\item{dt}{data for computation.}

\item{window}{The selected time window}
}
\description{
Generate 1d datatable summarizing 2d variables
reduce allows to reduce datatable according to time-window selected and according to
the iteration of values for selected variables, into a 1d table with one row by admission
}

