% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_apache_aki.R
\name{gen_apache_aki}
\alias{gen_apache_aki}
\title{Generates the APACHE Acute Kidney Injury score}
\usage{
gen_apache_aki(dt, crrt, window = c(0, 24))
}
\arguments{
\item{dt}{data.table containing physiological data}

\item{crrt}{Strings. Name of the Binary vector containing the assessment of prior chronic renal insufficiency.}

\item{window}{Numeric.Vector delimiting boundaries for time-window.}
}
\examples{
ddata <- NULL
ddata$"time" <- sample(seq(1,72,1), 200, replace = T)
ddata <- as.data.table(ddata)
ddata[, ("site") := sample(c("XX", "ZZ", "YY"), 200, replace = T)]
ddata[, ("episode_id") := sample(seq(1,250,1), 200, replace = T)]
ddata[, ("Creatinine") := sample(seq(20,1000,11), 200, replace = T)]
ddata[, crrt := sample(c(0,1), 200, replace = T)]
system.time(gen_apache_aki(ddata, crrt = ddata[, crrt], window = c(0,24), format = "dataItem"))
ddata[time \%between\% c(0,24), .N, by = c("site","episode_id", "apache_aki")]

}

