% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_haemo}
\alias{gen_haemo}
\title{Generate Haematocrit variable from the haemoglobin}
\usage{
gen_haemo(dt, mcch = NULL, format = "dataItem")
}
\arguments{
\item{dt}{data for computation.}

\item{mcch}{Numeric. An average value of mean corpuscular concentration of haemoglobin. Default is MCCH filled in datatable}

\item{format}{String. The format chosen for data items. Could be "dataItem", "shortName" or "NHICcode".
See relabel_cols for more informations.}
}
\description{
gen_grad derives a Haematocrit from haemoglobin and mean globular volume
}
\examples{
ddata <- NULL
ddata$"Haemoglobin" <- sample(seq(5, 25, 1), 200, replace = T)
ddata <- as.data.table(ddata)
gen_haemo(ddata, mcch = NULL, format = "dataItem")

}

