% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_q_rr}
\alias{gen_q_rr}
\title{Generate q_rr from respiratory fields}
\usage{
gen_q_rr(dt, qual = c(3, 60))
}
\arguments{
\item{dt}{data for computation.}

\item{qual}{vector of inner and outer bounds of quality assessment for respiratory rate.}
}
\description{
Generate q_rr from respiratory fields
}
\examples{
ddata <- NULL
ddata$"Spontaneous Respiratory Rate" <- sample(seq(0,100,1), 200, replace = T)
ddata <- as.data.table(ddata)
gen_q_rr(ddata, qual = c(3,60))

}

